Radeon 5650 (by slice). Refactoring

+++ FintekF718x/FintekF718x.h
@@ -93,9 +93,9 @@ private:
 	virtual void			enter();
 	virtual void			exit();
 	
-	virtual long			readTemperature(unsigned long index);
-	virtual long			readVoltage(unsigned long index);
-	virtual long			readTachometer(unsigned long index);
+	virtual UInt16			readTemperature(UInt32 index);
+	virtual UInt16			readVoltage(UInt32 index);
+	virtual UInt16			readTachometer(UInt32 index);
 	
 	virtual const char *	getModelName();
